chore: separate videooutputpip file
