+++ media_kit_video/macos/Classes/plugin/common/VideoOutputPIP.swift
@@ -1,0 +1,1 @@
+../../../../common/darwin/Classes/plugin/VideoOutputPIP.swift
